refactor(rust-driver): extract gRPC impl behind PactPlugin trait
Introduces a PactPlugin trait so the driver can support plugin
interface types beyond gRPC (Lua, Python, JS embedded interpreters)
without modifying existing code paths.

- New grpc_plugin.rs: GrpcPactPlugin struct (renamed from PactPlugin),
  PluginClient, PactPluginInterceptor, start_plugin_process — all
  gRPC-specific code in one place.
- plugin_models.rs: PactPlugin becomes a trait (operational async
  methods + synchronous accessors). PactPluginRpc slimmed to
  init_plugin only. Test doubles moved here as pub(crate).
- plugin_manager.rs: PLUGIN_REGISTER / load_plugin / lookup_plugin use
  Arc<dyn PactPlugin + Send + Sync>. Inner verify/prepare functions
  take &dyn PactPlugin (manifest from plugin.manifest()).
- mock_server.rs: MockServerDetails.plugin is Arc<dyn PactPlugin>.
- lib.rs: pub mod grpc_plugin added.

All 14 unit tests pass; no clippy errors.
