+++ drivers/rust/driver/src/plugin_manager.rs
@@ -1,25 +1,20 @@
 //! Manages interactions with Pact plugins
 use std::collections::HashMap;
 use std::env;
-use uuid::Uuid;
 use std::fs;
 use std::fs::File;
 use std::io::{BufReader, Write};
 use std::path::PathBuf;
-use std::process::Command;
-use std::process::Stdio;
 use std::str::FromStr;
 use std::str::from_utf8;
-use std::sync::Mutex;
+use std::sync::{Arc, Mutex};
 use std::thread;
 
 use anyhow::{Context, anyhow, bail};
 use bytes::Bytes;
 use itertools::Either;
 use lazy_static::lazy_static;
-use log::max_level;
 use maplit::hashmap;
-use os_info::Type;
 use pact_models::PactSpecification;
 use pact_models::bodies::OptionalBody;
 use pact_models::json_utils::json_to_string;
@@ -29,13 +24,11 @@ use pact_models::v4::interaction::V4Interaction;
 use reqwest::Client;
 use semver::Version;
 use serde_json::Value;
-use sysinfo::{Pid, System};
 use tracing::{debug, info, trace, warn};
 
 use crate::catalogue_manager::{
   CatalogueEntry, all_entries, core_entries, register_plugin_entries, remove_plugin_entries,
 };
-use crate::child_process::ChildPluginProcess;
 use crate::content::ContentMismatch;
 use crate::download::{download_json_from_github, download_plugin_executable, fetch_json_from_url};
 use crate::metrics::send_metrics;
@@ -55,7 +48,8 @@ use crate::verification::{InteractionVerificationData, InteractionVerificationRe
 lazy_static! {
   static ref PLUGIN_MANIFEST_REGISTER: Mutex<HashMap<String, PactPluginManifest>> =
     Mutex::new(HashMap::new());
-  static ref PLUGIN_REGISTER: Mutex<HashMap<String, PactPlugin>> = Mutex::new(HashMap::new());
+  static ref PLUGIN_REGISTER: Mutex<HashMap<String, Arc<dyn PactPlugin + Send + Sync>>> =
+    Mutex::new(HashMap::new());
   /// Maps plugin_instance_id → plugin_name so the PluginHost Log RPC handler can
   /// attach the plugin name to forwarded log entries without a proto field for it.
   static ref INSTANCE_NAMES: Mutex<HashMap<String, String>> = Mutex::new(HashMap::new());
@@ -65,11 +59,11 @@ pub(crate) fn plugin_name_for_instance(instance_id: &str) -> Option<String> {
   INSTANCE_NAMES.lock().unwrap().get(instance_id).cloned()
 }
 
-fn register_plugin_instance(instance_id: &str, plugin_name: &str) {
+pub(crate) fn register_plugin_instance(instance_id: &str, plugin_name: &str) {
   INSTANCE_NAMES.lock().unwrap().insert(instance_id.to_string(), plugin_name.to_string());
 }
 
-fn deregister_plugin_instance(instance_id: &str) {
+pub(crate) fn deregister_plugin_instance(instance_id: &str) {
   INSTANCE_NAMES.lock().unwrap().remove(instance_id);
 }
 
@@ -82,7 +76,7 @@ fn host_capabilities() -> Vec<String> {
 
 /// Load the plugin defined by the dependency information. Will first look in the global
 /// plugin registry.
-pub async fn load_plugin(plugin: &PluginDependency) -> anyhow::Result<PactPlugin> {
+pub async fn load_plugin(plugin: &PluginDependency) -> anyhow::Result<Arc<dyn PactPlugin + Send + Sync>> {
   let thread_id = thread::current().id();
   debug!("Loading plugin {:?}", plugin);
   trace!(
@@ -95,11 +89,11 @@ pub async fn load_plugin(plugin: &PluginDependency) -> anyhow::Result<PactPlugin
   );
   let mut inner = PLUGIN_REGISTER.lock().unwrap();
   trace!("load_plugin {:?}: Got PLUGIN_REGISTER lock", thread_id);
-  let result = match lookup_plugin_inner(plugin, &mut inner) {
-    Some(plugin) => {
-      debug!("Found running plugin {:?}", plugin);
-      plugin.update_access();
-      Ok(plugin.clone())
+  let result = match lookup_plugin_inner(plugin, &inner) {
+    Some(plugin_arc) => {
+      debug!("Found running plugin {:?}", plugin_arc);
+      plugin_arc.update_access();
+      Ok(plugin_arc)
     }
     None => {
       debug!("Did not find plugin, will attempt to start it");
@@ -134,36 +128,36 @@ pub async fn load_plugin(plugin: &PluginDependency) -> anyhow::Result<PactPlugin
   result
 }
 
-fn lookup_plugin_inner<'a>(
+fn lookup_plugin_inner(
   plugin: &PluginDependency,
-  plugin_register: &'a mut HashMap<String, PactPlugin>,
-) -> Option<&'a mut PactPlugin> {
+  plugin_register: &HashMap<String, Arc<dyn PactPlugin + Send + Sync>>,
+) -> Option<Arc<dyn PactPlugin + Send + Sync>> {
   if let Some(version) = &plugin.version {
-    plugin_register.get_mut(format!("{}/{}", plugin.name, version).as_str())
+    plugin_register.get(format!("{}/{}", plugin.name, version).as_str()).cloned()
   } else {
     plugin_register
-      .iter_mut()
-      .filter(|(_, value)| value.manifest.name == plugin.name)
-      .max_by(|(_, v1), (_, v2)| v1.manifest.version.cmp(&v2.manifest.version))
-      .map(|(_, plugin)| plugin)
+      .iter()
+      .filter(|(_, value)| value.manifest().name == plugin.name)
+      .max_by(|(_, v1), (_, v2)| v1.manifest().version.cmp(&v2.manifest().version))
+      .map(|(_, plugin)| plugin.clone())
   }
 }
 
 /// Look up the plugin in the global plugin register
-pub fn lookup_plugin(plugin: &PluginDependency) -> Option<PactPlugin> {
+pub fn lookup_plugin(plugin: &PluginDependency) -> Option<Arc<dyn PactPlugin + Send + Sync>> {
   let thread_id = thread::current().id();
   trace!(
     "lookup_plugin {:?}: Waiting on PLUGIN_REGISTER lock",
     thread_id
   );
-  let mut inner = PLUGIN_REGISTER.lock().unwrap();
+  let inner = PLUGIN_REGISTER.lock().unwrap();
   trace!("lookup_plugin {:?}: Got PLUGIN_REGISTER lock", thread_id);
-  let entry = lookup_plugin_inner(plugin, &mut inner);
+  let entry = lookup_plugin_inner(plugin, &inner);
   trace!(
     "lookup_plugin {:?}: Releasing PLUGIN_REGISTER lock",
     thread_id
   );
-  entry.cloned()
+  entry
 }
 
 /// Return the plugin manifest for the given plugin. Will first look in the global plugin manifest
@@ -269,8 +263,8 @@ pub fn lookup_plugin_manifest(plugin: &PluginDependency) -> Option<PactPluginMan
 
 async fn initialise_plugin(
   manifest: &PactPluginManifest,
-  plugin_register: &mut HashMap<String, PactPlugin>,
-) -> anyhow::Result<PactPlugin> {
+  plugin_register: &mut HashMap<String, Arc<dyn PactPlugin + Send + Sync>>,
+) -> anyhow::Result<Arc<dyn PactPlugin + Send + Sync>> {
   let interface_version = PluginInterfaceVersion::try_from(manifest.plugin_interface_version)
     .with_context(|| {
       format!(
@@ -283,13 +277,13 @@ async fn initialise_plugin(
     PluginInterfaceVersion::V1 | PluginInterfaceVersion::V2 => {
       match manifest.executable_type.as_str() {
         "exec" => {
-          let mut plugin = start_plugin_process(manifest).await?;
+          let mut plugin = crate::grpc_plugin::start_plugin_process(manifest).await?;
           debug!(
             "Plugin process started OK (port = {}), sending init message",
             plugin.port()
           );
 
-          let instance_id = plugin.instance_id.clone();
+          let instance_id = plugin.instance_id().to_string();
           let response = init_handshake(manifest, &mut plugin, &instance_id).await.map_err(|err| {
             deregister_plugin_instance(&instance_id);
             plugin.kill();
@@ -297,6 +291,7 @@ async fn initialise_plugin(
           })?;
           plugin.plugin_capabilities = response.plugin_capabilities;
 
+          let plugin: Arc<dyn PactPlugin + Send + Sync> = Arc::new(plugin);
           let key = format!("{}/{}", manifest.name, manifest.version);
           plugin_register.insert(key, plugin.clone());
 
@@ -333,92 +328,6 @@ pub async fn init_handshake(
   Ok(response)
 }
 
-async fn start_plugin_process(manifest: &PactPluginManifest) -> anyhow::Result<PactPlugin> {
-  debug!("Starting plugin with manifest {:?}", manifest);
-
-  let os_info = os_info::get();
-  debug!("Detected OS: {}", os_info);
-  let mut path = if let Some(entry_point) = manifest.entry_points.get(&os_info.to_string()) {
-    PathBuf::from(entry_point)
-  } else if os_info.os_type() == Type::Windows && manifest.entry_points.contains_key("windows") {
-    PathBuf::from(manifest.entry_points.get("windows").unwrap())
-  } else {
-    PathBuf::from(&manifest.entry_point)
-  };
-  if !path.is_absolute() || !path.exists() {
-    path = PathBuf::from(manifest.plugin_dir.clone()).join(path);
-  }
-  debug!("Starting plugin using {:?}", &path);
-
-  let host_port = match crate::plugin_host::ensure_plugin_host_running().await {
-    Ok(port) => Some(port),
-    Err(err) => {
-      warn!("Could not start PluginHost server, Log RPC forwarding will be unavailable: {}", err);
-      None
-    }
-  };
-
-  let log_level = max_level();
-  let mut child_command = Command::new(path.clone());
-  let mut child_command = child_command
-    .env("LOG_LEVEL", log_level.to_string())
-    .env("RUST_LOG", log_level.to_string())
-    .current_dir(manifest.plugin_dir.clone());
-
-  let instance_id = Uuid::new_v4().to_string();
-
-  child_command = child_command.env("PACT_PLUGIN_INSTANCE_ID", &instance_id);
-  if let Some(port) = host_port {
-    child_command = child_command.env("PACT_PLUGIN_HOST", format!("127.0.0.1:{}", port));
-  }
-
-  if let Some(args) = &manifest.args {
-    child_command = child_command.args(args);
-  }
-
-  let child = child_command
-    .stdout(Stdio::piped())
-    .stderr(Stdio::piped())
-    .spawn()
-    .map_err(|err| {
-      anyhow!(
-        "Was not able to start plugin process for '{}' - {}",
-        path.to_string_lossy(),
-        err
-      )
-    })?;
-  let child_pid = child.id();
-  debug!("Plugin {} started with PID {} (instance {})", manifest.name, child_pid, instance_id);
-  register_plugin_instance(&instance_id, &manifest.name);
-
-  match ChildPluginProcess::new(child, manifest, instance_id.clone()).await {
-    Ok(child) => {
-      let plugin = PactPlugin::new(manifest, child)?;
-      Ok(plugin)
-    }
-    Err(err) => {
-      deregister_plugin_instance(&instance_id);
-      let mut s = System::new();
-      s.refresh_processes();
-      if let Some(process) = s.process(Pid::from_u32(child_pid)) {
-        #[cfg(not(windows))]
-        process.kill();
-        // revert windows specific logic once https://github.com/GuillaumeGomez/sysinfo/pull/1341/files is merged/released
-        #[cfg(windows)]
-        let _ = Command::new("taskkill.exe")
-          .arg("/PID")
-          .arg(child_pid.to_string())
-          .arg("/F")
-          .arg("/T")
-          .output();
-      } else {
-        warn!("Child process with PID {} was not found", child_pid);
-      }
-      Err(err)
-    }
-  }
-}
-
 /// Shut down all plugin processes
 pub fn shutdown_plugins() {
   let thread_id = thread::current().id();
@@ -431,9 +340,9 @@ pub fn shutdown_plugins() {
   trace!("shutdown_plugins {:?}: Got PLUGIN_REGISTER lock", thread_id);
   for plugin in guard.values() {
     debug!("Shutting down plugin {:?}", plugin);
-    deregister_plugin_instance(&plugin.instance_id);
+    deregister_plugin_instance(plugin.instance_id());
     plugin.kill();
-    remove_plugin_entries(&plugin.manifest.name);
+    remove_plugin_entries(&plugin.manifest().name);
   }
   guard.clear();
   trace!(
@@ -443,14 +352,14 @@ pub fn shutdown_plugins() {
 }
 
 /// Shutdown the given plugin
-pub fn shutdown_plugin(plugin: &mut PactPlugin) {
+pub fn shutdown_plugin(plugin: &dyn PactPlugin) {
   debug!(
     "Shutting down plugin {}:{}",
-    plugin.manifest.name, plugin.manifest.version
+    plugin.manifest().name, plugin.manifest().version
   );
-  deregister_plugin_instance(&plugin.instance_id);
+  deregister_plugin_instance(plugin.instance_id());
   plugin.kill();
-  remove_plugin_entries(&plugin.manifest.name);
+  remove_plugin_entries(&plugin.manifest().name);
 }
 
 /// Publish the current catalogue to all plugins
@@ -490,7 +399,7 @@ pub async fn publish_updated_catalogue() {
     if let Err(err) = plugin.update_catalogue(request.clone()).await {
       warn!(
         "Failed to send updated catalogue to plugin '{}' - {}",
-        plugin.manifest.name, err
+        plugin.manifest().name, err
       );
     }
   }
@@ -505,14 +414,14 @@ pub fn increment_plugin_access(plugin: &PluginDependency) {
     "increment_plugin_access {:?}: Waiting on PLUGIN_REGISTER lock",
     thread_id
   );
-  let mut inner = PLUGIN_REGISTER.lock().unwrap();
+  let inner = PLUGIN_REGISTER.lock().unwrap();
   trace!(
     "increment_plugin_access {:?}: Got PLUGIN_REGISTER lock",
     thread_id
   );
 
-  if let Some(plugin) = lookup_plugin_inner(plugin, &mut inner) {
-    plugin.update_access();
+  if let Some(plugin_arc) = lookup_plugin_inner(plugin, &inner) {
+    plugin_arc.update_access();
   }
 
   trace!(
@@ -531,15 +440,16 @@ pub fn drop_plugin_access(plugin: &PluginDependency) {
     thread_id
   );
   let mut inner = PLUGIN_REGISTER.lock().unwrap();
-  trace!(
-    "drop_plugin_access {:?}: Got PLUGIN_REGISTER lock",
-    thread_id
-  );
+  trace!("drop_plugin_access {:?}: Got PLUGIN_REGISTER lock", thread_id);
 
-  if let Some(plugin) = lookup_plugin_inner(plugin, &mut inner) {
-    let key = format!("{}/{}", plugin.manifest.name, plugin.manifest.version);
-    if plugin.drop_access() == 0 {
-      shutdown_plugin(plugin);
+  if let Some(plugin_arc) = lookup_plugin_inner(plugin, &inner) {
+    let key = format!(
+      "{}/{}",
+      plugin_arc.manifest().name,
+      plugin_arc.manifest().version
+    );
+    if plugin_arc.drop_access() == 0 {
+      shutdown_plugin(plugin_arc.as_ref());
       inner.remove(key.as_str());
     }
   }
@@ -717,8 +627,8 @@ pub async fn shutdown_mock_server(
   };
 
   debug!(
-    plugin_name = mock_server.plugin.manifest.name.as_str(),
-    plugin_version = mock_server.plugin.manifest.version.as_str(),
+    plugin_name = mock_server.plugin.manifest().name.as_str(),
+    plugin_version = mock_server.plugin.manifest().version.as_str(),
     server_key = mock_server.key.as_str(),
     "Sending shutdownMockServer request to plugin"
   );
@@ -770,8 +680,8 @@ pub async fn get_mock_server_results(
   };
 
   debug!(
-    plugin_name = mock_server.plugin.manifest.name.as_str(),
-    plugin_version = mock_server.plugin.manifest.version.as_str(),
+    plugin_name = mock_server.plugin.manifest().name.as_str(),
+    plugin_version = mock_server.plugin.manifest().version.as_str(),
     server_key = mock_server.key.as_str(),
     "Sending getMockServerResults request to plugin"
   );
@@ -828,16 +738,16 @@ pub async fn prepare_validation_for_interaction(
   let plugin = lookup_plugin(&manifest.as_dependency())
     .ok_or_else(|| anyhow!("Did not find a running plugin for manifest {:?}", manifest))?;
 
-  prepare_validation_for_interaction_inner(&plugin, manifest, pact, interaction, context).await
+  prepare_validation_for_interaction_inner(plugin.as_ref(), pact, interaction, context).await
 }
 
 pub(crate) async fn prepare_validation_for_interaction_inner(
-  plugin: &(dyn PactPluginRpc + Send + Sync),
-  manifest: &PactPluginManifest,
+  plugin: &dyn PactPlugin,
   pact: &V4Pact,
   interaction: &(dyn V4Interaction + Send + Sync),
   context: &HashMap<String, Value>,
 ) -> anyhow::Result<InteractionVerificationData> {
+  let manifest = plugin.manifest();
   debug!(
     plugin_name = manifest.name.as_str(),
     plugin_version = manifest.version.as_str(),
@@ -927,8 +837,7 @@ pub async fn verify_interaction(
     .ok_or_else(|| anyhow!("Did not find a running plugin for manifest {:?}", manifest))?;
 
   verify_interaction_inner(
-    &plugin,
-    &manifest,
+    plugin.as_ref(),
     verification_data,
     config,
     pact,
@@ -938,13 +847,13 @@ pub async fn verify_interaction(
 }
 
 pub(crate) async fn verify_interaction_inner(
-  plugin: &(dyn PactPluginRpc + Send + Sync),
-  manifest: &PactPluginManifest,
+  plugin: &dyn PactPlugin,
   verification_data: &InteractionVerificationData,
   config: &HashMap<String, Value>,
   pact: &V4Pact,
   interaction: &(dyn V4Interaction + Send + Sync),
 ) -> anyhow::Result<InteractionVerificationResult> {
+  let manifest = plugin.manifest();
   debug!(
     plugin_name = manifest.name.as_str(),
     plugin_version = manifest.version.as_str(),
@@ -1083,189 +992,25 @@ fn create_plugin_dir(manifest: &PactPluginManifest) -> anyhow::Result<PathBuf> {
 mod tests {
   use std::collections::HashMap;
   use std::fs::{self, File};
-  use std::sync::RwLock;
+  use std::sync::Arc;
 
-  use async_trait::async_trait;
+  use expectest::prelude::*;
   use maplit::hashmap;
   use pact_models::prelude::v4::V4Pact;
   use pact_models::v4::interaction::V4Interaction;
   use pact_models::v4::sync_message::SynchronousMessage;
-
-  use expectest::prelude::*;
   use tempdir::TempDir;
-
-  use crate::plugin_manager::prepare_validation_for_interaction_inner;
-  use crate::plugin_manager::verify_interaction_inner;
-  use crate::plugin_models::PluginDependency;
-  use crate::plugin_models::tests::MockPlugin;
-  use crate::proto::*;
-  use crate::verification::InteractionVerificationData;
 
   use crate::catalogue_manager::{
     CatalogueEntry, CatalogueEntryProviderType, CatalogueEntryType, register_core_entries,
   };
-
-  use super::{PactPluginManifest, init_handshake, initialise_plugin, load_manifest_from_dir};
-
-  struct FailingInitPlugin {
-    error: String,
-  }
-
-  #[async_trait]
-  impl crate::plugin_models::PactPluginRpc for FailingInitPlugin {
-    async fn init_plugin(
-      &mut self,
-      _request: crate::plugin_models::PluginInitRequest,
-    ) -> anyhow::Result<crate::plugin_models::PluginInitResponse> {
-      Err(anyhow::anyhow!("{}", self.error))
-    }
-
-    async fn compare_contents(
-      &self,
-      _request: CompareContentsRequest,
-    ) -> anyhow::Result<CompareContentsResponse> {
-      unimplemented!()
-    }
-
-    async fn configure_interaction(
-      &self,
-      _request: ConfigureInteractionRequest,
-    ) -> anyhow::Result<ConfigureInteractionResponse> {
-      unimplemented!()
-    }
-
-    async fn generate_content(
-      &self,
-      _request: GenerateContentRequest,
-    ) -> anyhow::Result<GenerateContentResponse> {
-      unimplemented!()
-    }
-
-    async fn start_mock_server(
-      &self,
-      _request: StartMockServerRequest,
-    ) -> anyhow::Result<StartMockServerResponse> {
-      unimplemented!()
-    }
-
-    async fn shutdown_mock_server(
-      &self,
-      _request: ShutdownMockServerRequest,
-    ) -> anyhow::Result<ShutdownMockServerResponse> {
-      unimplemented!()
-    }
-
-    async fn get_mock_server_results(
-      &self,
-      _request: MockServerRequest,
-    ) -> anyhow::Result<MockServerResults> {
-      unimplemented!()
-    }
-
-    async fn prepare_interaction_for_verification(
-      &self,
-      _request: VerificationPreparationRequest,
-    ) -> anyhow::Result<VerificationPreparationResponse> {
-      unimplemented!()
-    }
-
-    async fn verify_interaction(
-      &self,
-      _request: VerifyInteractionRequest,
-    ) -> anyhow::Result<VerifyInteractionResponse> {
-      unimplemented!()
-    }
-
-    async fn update_catalogue(&self, _request: Catalogue) -> anyhow::Result<()> {
-      unimplemented!()
-    }
-  }
-
-  struct InitRecordingPlugin {
-    request: RwLock<Option<crate::plugin_models::PluginInitRequest>>,
-  }
-
-  impl Default for InitRecordingPlugin {
-    fn default() -> Self {
-      Self {
-        request: RwLock::new(None),
-      }
-    }
-  }
-
-  #[async_trait]
-  impl crate::plugin_models::PactPluginRpc for InitRecordingPlugin {
-    async fn init_plugin(
-      &mut self,
-      request: crate::plugin_models::PluginInitRequest,
-    ) -> anyhow::Result<crate::plugin_models::PluginInitResponse> {
-      *self.request.write().unwrap() = Some(request);
-      Ok(crate::plugin_models::PluginInitResponse {
-        catalogue: vec![],
-        plugin_capabilities: vec!["interaction/request-response".to_string()],
-      })
-    }
-
-    async fn compare_contents(
-      &self,
-      _request: CompareContentsRequest,
-    ) -> anyhow::Result<CompareContentsResponse> {
-      unimplemented!()
-    }
-
-    async fn configure_interaction(
-      &self,
-      _request: ConfigureInteractionRequest,
-    ) -> anyhow::Result<ConfigureInteractionResponse> {
-      unimplemented!()
-    }
-
-    async fn generate_content(
-      &self,
-      _request: GenerateContentRequest,
-    ) -> anyhow::Result<GenerateContentResponse> {
-      unimplemented!()
-    }
-
-    async fn start_mock_server(
-      &self,
-      _request: StartMockServerRequest,
-    ) -> anyhow::Result<StartMockServerResponse> {
-      unimplemented!()
-    }
-
-    async fn shutdown_mock_server(
-      &self,
-      _request: ShutdownMockServerRequest,
-    ) -> anyhow::Result<ShutdownMockServerResponse> {
-      unimplemented!()
-    }
-
-    async fn get_mock_server_results(
-      &self,
-      _request: MockServerRequest,
-    ) -> anyhow::Result<MockServerResults> {
-      unimplemented!()
-    }
-
-    async fn prepare_interaction_for_verification(
-      &self,
-      _request: VerificationPreparationRequest,
-    ) -> anyhow::Result<VerificationPreparationResponse> {
-      unimplemented!()
-    }
-
-    async fn verify_interaction(
-      &self,
-      _request: VerifyInteractionRequest,
-    ) -> anyhow::Result<VerifyInteractionResponse> {
-      unimplemented!()
-    }
-
-    async fn update_catalogue(&self, _request: Catalogue) -> anyhow::Result<()> {
-      Ok(())
-    }
-  }
+  use crate::plugin_manager::{
+    init_handshake, initialise_plugin, load_manifest_from_dir,
+    prepare_validation_for_interaction_inner, verify_interaction_inner,
+  };
+  use crate::plugin_models::tests::{FailingInitPlugin, InitRecordingPlugin, MockPlugin};
+  use crate::plugin_models::{PactPlugin, PactPluginManifest, PluginDependency};
+  use crate::verification::InteractionVerificationData;
 
   #[test]
   fn load_manifest_from_dir_test() {
@@ -1341,7 +1086,7 @@ mod tests {
       ..PactPluginManifest::default()
     };
 
-    let mut plugin_register = HashMap::new();
+    let mut plugin_register: HashMap<String, Arc<dyn PactPlugin + Send + Sync>> = HashMap::new();
     let err = initialise_plugin(&manifest, &mut plugin_register)
       .await
       .unwrap_err();
@@ -1399,12 +1144,14 @@ mod tests {
 
   #[test_log::test(tokio::test)]
   async fn prepare_validation_for_interaction_passes_in_pact_with_interaction_keys_set() {
-    let manifest = PactPluginManifest {
-      name: "test-plugin".to_string(),
-      version: "0.0.0".to_string(),
-      ..PactPluginManifest::default()
+    let mock_plugin = MockPlugin {
+      manifest: PactPluginManifest {
+        name: "test-plugin".to_string(),
+        version: "0.0.0".to_string(),
+        ..PactPluginManifest::default()
+      },
+      ..MockPlugin::default()
     };
-    let mock_plugin = MockPlugin::default();
 
     let interaction = SynchronousMessage {
       ..SynchronousMessage::default()
@@ -1417,7 +1164,6 @@ mod tests {
 
     let result = prepare_validation_for_interaction_inner(
       &mock_plugin,
-      &manifest,
       &pact,
       &interaction,
       &context,
@@ -1436,12 +1182,14 @@ mod tests {
 
   #[test_log::test(tokio::test)]
   async fn prepare_validation_for_interaction_handles_pact_with_keys_already_set() {
-    let manifest = PactPluginManifest {
-      name: "test-plugin".to_string(),
-      version: "0.0.0".to_string(),
-      ..PactPluginManifest::default()
+    let mock_plugin = MockPlugin {
+      manifest: PactPluginManifest {
+        name: "test-plugin".to_string(),
+        version: "0.0.0".to_string(),
+        ..PactPluginManifest::default()
+      },
+      ..MockPlugin::default()
     };
-    let mock_plugin = MockPlugin::default();
 
     let interaction = SynchronousMessage {
       key: Some("1234567890".to_string()),
@@ -1455,7 +1203,6 @@ mod tests {
 
     let result = prepare_validation_for_interaction_inner(
       &mock_plugin,
-      &manifest,
       &pact,
       &interaction,
       &context,
@@ -1475,12 +1222,14 @@ mod tests {
 
   #[test_log::test(tokio::test)]
   async fn verify_interaction_passes_in_pact_with_interaction_keys_set() {
-    let manifest = PactPluginManifest {
-      name: "test-plugin".to_string(),
-      version: "0.0.0".to_string(),
-      ..PactPluginManifest::default()
+    let mock_plugin = MockPlugin {
+      manifest: PactPluginManifest {
+        name: "test-plugin".to_string(),
+        version: "0.0.0".to_string(),
+        ..PactPluginManifest::default()
+      },
+      ..MockPlugin::default()
     };
-    let mock_plugin = MockPlugin::default();
 
     let interaction = SynchronousMessage {
       ..SynchronousMessage::default()
@@ -1494,7 +1243,6 @@ mod tests {
 
     let result = verify_interaction_inner(
       &mock_plugin,
-      &manifest,
       &data,
       &context,
       &pact,
@@ -1514,12 +1262,14 @@ mod tests {
 
   #[test_log::test(tokio::test)]
   async fn verify_interaction_handles_interaction_with_key_already_set() {
-    let manifest = PactPluginManifest {
-      name: "test-plugin".to_string(),
-      version: "0.0.0".to_string(),
-      ..PactPluginManifest::default()
+    let mock_plugin = MockPlugin {
+      manifest: PactPluginManifest {
+        name: "test-plugin".to_string(),
+        version: "0.0.0".to_string(),
+        ..PactPluginManifest::default()
+      },
+      ..MockPlugin::default()
     };
-    let mock_plugin = MockPlugin::default();
 
     let interaction = SynchronousMessage {
       key: Some("1234567890".to_string()),
@@ -1534,7 +1284,6 @@ mod tests {
 
     let result = verify_interaction_inner(
       &mock_plugin,
-      &manifest,
       &data,
       &context,
       &pact,
